ENH: Limit handle_data to market times.
To prevent cases where custom data types had unaligned timestamps,
check the event dt against the market_open and only call handle_data
if the event is at the market open or later.

The data bar updates with the non-market window events, as it did
before. When handle_data is called at market open, the last seen
value will be included in data.